orchestrator/requirements.txt
# Cubiczan Orchestrator Dependencies
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
redis>=5.0.0
neo4j>=5.0.0
qdrant-client>=1.7.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=6.0.0
sentence-transformers>=3.0.0
